contributions/core:  - Moved get_image function out of contributions/tests/media/model_factories.py to core/tests/helpers/image_helpers.py so it can be more easily used in different places while avoiding import cycles
categories:
 - Category model: symbol field path now has max_length 500
 - LookupValue mpdel: added symbol field: this is a first step towards supporting a field-value-based symbology mode in Category
 - model_factories.py: use get_image to generate symbol PNGs for LookupValues
 - test_manager.py, test_model.py & test_views.py: proper clean-up of all generated symbol image files in tearDown() methods
