+++ geokey/categories/tests/test_manager.py
@@ -443,6 +443,13 @@ class FieldManagerTest(TestCase):
 
 
 class LookupManagerTest(TestCase):
+    def tearDown(self):
+        for lookup_value in LookupValue.objects.all():
+            try:
+                lookup_value.symbol.delete()
+            except BaseException:
+                pass
+
     def test(self):
         LookupValueFactory.create_batch(5, **{'status': 'active'})
         LookupValueFactory.create_batch(5, **{'status': 'deleted'})
