+++ geokey/categories/tests/test_model.py
@@ -1,11 +1,12 @@
 import json
+
 from django.test import TestCase
 
 from nose.tools import raises
 
 from geokey.core.exceptions import InputError
 
-from ..models import Field, Category
+from ..models import Field, Category, LookupValue
 
 from .model_factories import (
     TextFieldFactory, NumericFieldFactory, DateTimeFieldFactory,
@@ -451,6 +452,13 @@ class NumericFieldTest(TestCase):
 
 
 class SingleLookupFieldTest(TestCase):
+    def tearDown(self):
+        for lookup_value in LookupValue.objects.all():
+            try:
+                lookup_value.symbol.delete()
+            except BaseException:
+                pass
+
     def test_create_lookupfield(self):
         category = CategoryFactory()
         field = Field.create(
