+++ geokey/contributions/tests/media/model_factories.py
@@ -1,24 +1,9 @@
 import factory
-from PIL import Image
-from StringIO import StringIO
-
-from django.core.files.base import ContentFile
 
 from geokey.users.tests.model_factories import UserFactory
 from geokey.contributions.tests.model_factories import ObservationFactory
 from geokey.contributions.models import ImageFile, VideoFile
-
-
-def get_image(file_name='test.png', width=200, height=200):
-    image_file = StringIO()
-    image = Image.new('RGBA', size=(width, height), color=(255, 0, 255))
-    image.save(image_file, 'png')
-    image_file.seek(0)
-
-    the_file = ContentFile(image_file.read(), file_name)
-    the_file.content_type = 'image/png'
-
-    return the_file
+from geokey.core.tests.helpers.image_helpers import get_image
 
 
 class ImageFileFactory(factory.django.DjangoModelFactory):
